+++ src/secrets/components/CreateSecretOverlay.tsx
@@ -1,0 +1,37 @@
+// Libraries
+import React, {FC} from 'react'
+import {useHistory} from 'react-router-dom'
+
+// Components
+import {Overlay} from '@influxdata/clockface'
+import CreateSecretForm from 'src/secrets/components/CreateSecretForm'
+
+// Types
+import ErrorBoundary from 'src/shared/components/ErrorBoundary'
+import GetResources from 'src/resources/components/GetResources'
+import {ResourceType} from 'src/types'
+
+const CreateSecretOverlay: FC = () => {
+  const history = useHistory()
+
+  return (
+    <Overlay visible={true}>
+      <Overlay.Container maxWidth={750}>
+        <Overlay.Header
+          title="Add Secret"
+          onDismiss={() => history.goBack()}
+          testID="modify-secret-overlay-header"
+        />
+        <Overlay.Body>
+          <GetResources resources={[ResourceType.Secrets]}>
+            <ErrorBoundary>
+              <CreateSecretForm />
+            </ErrorBoundary>
+          </GetResources>
+        </Overlay.Body>
+      </Overlay.Container>
+    </Overlay>
+  )
+}
+
+export default CreateSecretOverlay
