+++ src/shared/selectors/flags.ts
@@ -96,6 +96,7 @@ export const CLOUD_FLAGS = {
   'molly-first': false,
   'managed-functions': false,
   simpleTable: false,
+  secretsUI: true,
 }
 
 export const activeFlags = (state: AppState): FlagMap => {
